Fix linting issues (pylint)
