+++ amuser/base.py
@@ -3,10 +3,6 @@
 import os
 import shutil
 
-from selenium.webdriver.common.by import By
-from selenium.webdriver.support import expected_conditions as EC
-from selenium.webdriver.support.ui import WebDriverWait
-
 from . import constants as c
 from . import urls
 from . import utils
@@ -90,19 +86,3 @@ class Base:
                 os.unlink(thing_path)
             elif os.path.isdir(thing_path):
                 shutil.rmtree(thing_path)
-
-    @property
-    def ss_api_key(self):
-        if not self._ss_api_key:
-            self.driver.get(self.get_ss_login_url())
-            self.driver.find_element_by_id('id_username').send_keys(self.ss_username)
-            self.driver.find_element_by_id('id_password').send_keys(self.ss_password)
-            self.driver.find_element_by_css_selector(
-                c.varvn('SELECTOR_SS_LOGIN_BUTTON', self.vn)).click()
-            self.driver.get(self.get_default_ss_user_edit_url())
-            block = WebDriverWait(self.driver, 20)
-            block.until(EC.presence_of_element_located(
-                (By.CSS_SELECTOR, 'code')))
-            self._ss_api_key = self.driver.find_element_by_tag_name(
-                'code').text.strip()
-        return self._ss_api_key
