+++ runs.json
@@ -1,17 +1,20 @@
-{
-    "features/core/metadata-only-aip-reingest.feature": {
+[
+    {
         "behave_command": "behave --tags=mo-aip-reingest --no-skipped -v -D am_username=test -D am_password=test -D am_url=http://127.0.0.1:62080/ -D am_version=1.7 -D am_api_key=test -D ss_username=test -D ss_password=test -D ss_url=http://127.0.0.1:62081/ -D ss_api_key=${AM_SS_API_KEY} -D home=archivematica -D driver_name=Firefox",
-        "environments": {
-            [
-                {
-                    "archivematica_version": "1.7.0",
-                    "archivematica_branch": "qa/1.x",
-                    "archivematica_commit": "924fbb397ad58d334c0495f7cec4ed872535a983",
-                    "storage_service_version": "0.11.0",
-                    "storage_service_branch": "qa/0.x",
-                    "storage_service_commit": "edc78d40340a153ddba879291364463c06e4e4a2"
-                }
-            ]
-        }
+        "environment": {
+            "archivematica_version": "1.7.0",
+            "archivematica_branch": "qa/1.x",
+            "archivematica_commit": "924fbb397ad58d334c0495f7cec4ed872535a983",
+            "storage_service_version": "0.11.0",
+            "storage_service_branch": "qa/0.x",
+            "storage_service_commit": "edc78d40340a153ddba879291364463c06e4e4a2",
+            "sampledata_branch": "master",
+            "sampledata_commit": "b51ad285b063f8b99cb87e1fdb8e681ee5021708",
+            "acceptance_tests_branch": "dev/refactor",
+            "acceptance_tests_commit": "11564d5ea13844144ae2e651283a0482f3776533"
+        },
+        "target_features": ["features/core/metadata-only-aip-reingest.feature"],
+        "outcome": "fail",
+        "outcome_note": "Failed at step ``And the user initiates a metadata-only re-ingest on the AIP``, module am_browser_ability.py, line 148, in ``initiate_reingest`` with error `selenium.common.exceptions.NoSuchElementException: Message: Unable to locate element: div.alert-success`"
     }
-}
+]
